@@ -79,8 +79,8 @@
     </build>
 
     <properties>
-        <spring-kafka.version>3.1.2</spring-kafka.version>
-        <spring-boot.version>3.2.2</spring-boot.version>
+        <spring-kafka.version>3.3.1</spring-kafka.version>
+        <spring-boot.version>3.4.1</spring-boot.version>
     </properties>
 
 </project>
