@@ -12,6 +12,6 @@ public class RawTypesUnitTest {
         List myList = new ArrayList();
         myList.add(new Object());
         myList.add("2");
-        myList.add(new Integer(1));
+        myList.add(Integer.valueOf(1));
     }
 }
