@@ -123,20 +123,20 @@ public class GuavaUnitTest {
         BiMap<Integer, Integer> map = HashBiMap.create();
         start = System.nanoTime();
         for (int i = 0; i < 100000; i++) {
-            Integer key = new Integer(i);
-            Integer value = new Integer(i + 1);
+            Integer key = Integer.valueOf(i);
+            Integer value = Integer.valueOf(i + 1);
             map.put(key, value);
         }
         System.out.println("Insertion time:" + TimeUnit.MILLISECONDS.convert(System.nanoTime() - start, TimeUnit.NANOSECONDS));
 
         start = System.nanoTime();
-        Integer value = map.get(new Integer(500));
+        Integer value = map.get(Integer.valueOf(500));
         System.out.println("Value:" + value);
         System.out.println("Fetch time key:" + TimeUnit.MICROSECONDS.convert(System.nanoTime() - start, TimeUnit.NANOSECONDS));
 
         start = System.nanoTime();
         Integer key = map.inverse()
-            .get(new Integer(501));
+            .get(Integer.valueOf(501));
         System.out.println("Key:" + key);
         System.out.println("Fetch time value:" + TimeUnit.MICROSECONDS.convert(System.nanoTime() - start, TimeUnit.NANOSECONDS));
     }
