@@ -5,73 +5,72 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.regex.Pattern;
 
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.slf4j.LoggerFactory;
 
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.LoggerContext;
+import org.springframework.boot.test.system.CapturedOutput;
+import org.springframework.boot.test.system.OutputCaptureExtension;
 
-public class BusinessWorkerUnitTest {
+@ExtendWith(OutputCaptureExtension.class)
+class BusinessWorkerUnitTest {
 
-    private static MemoryAppender memoryAppender;
+    private final MemoryAppender memoryAppender = new MemoryAppender();
     private static final String LOGGER_NAME = "com.baeldung.junit.log";
     private static final String MSG = "This is a test message!!!";
 
-    @Before
-    public void setup() {
+    @BeforeEach
+    void setup() {
         Logger logger = (Logger) LoggerFactory.getLogger(LOGGER_NAME);
-        memoryAppender = new MemoryAppender();
-        memoryAppender.setContext((LoggerContext) LoggerFactory.getILoggerFactory());
         logger.setLevel(Level.DEBUG);
         logger.addAppender(memoryAppender);
-        memoryAppender.start();
 
+        memoryAppender.setContext((LoggerContext) LoggerFactory.getILoggerFactory());
+        memoryAppender.start();
     }
 
-    @After
-    public void cleanUp() {
+    @AfterEach
+    void cleanUp() {
         memoryAppender.reset();
         memoryAppender.stop();
     }
 
     @Test
-    public void test() {
+    void whenLogsGenerated_thenMemoryAppenderContainsLogs() {
         BusinessWorker worker = new BusinessWorker();
         worker.generateLogs(MSG);
 
-        // I check that I only have 4 messages (all but trace)
         assertThat(memoryAppender.countEventsForLogger(LOGGER_NAME)).isEqualTo(4);
-        // I look for a specific message at a specific level, and I only have 1
-        assertThat(memoryAppender.search(MSG, Level.INFO)
-            .size()).isEqualTo(1);
-        // I check that the entry that is not present is the trace level
+        assertThat(memoryAppender.search(MSG, Level.INFO)).hasSize(1);
         assertThat(memoryAppender.contains(MSG, Level.TRACE)).isFalse();
     }
 
     @Test
-    public void whenMultipleLogLevel_thenReturnExpectedResult() {
+    void whenMultipleLogLevel_thenReturnExpectedResult() {
         BusinessWorker worker = new BusinessWorker();
         worker.generateLogs("Transaction started for Order ID: 1001");
+
         assertThat(memoryAppender.countEventsForLogger(LOGGER_NAME)).isEqualTo(4);
-        assertThat(memoryAppender.search("Transaction started", Level.INFO)
-            .size()).isEqualTo(1);
-        assertThat(memoryAppender.search("Transaction started", Level.WARN)
-            .size()).isEqualTo(1);
-        assertThat(memoryAppender.search("Transaction started", Level.ERROR)
-            .size()).isEqualTo(1);
+        assertThat(memoryAppender.search("Transaction started", Level.DEBUG)).hasSize(1);
+        assertThat(memoryAppender.search("Transaction started", Level.INFO)).hasSize(1);
+        assertThat(memoryAppender.search("Transaction started", Level.WARN)).hasSize(1);
+        assertThat(memoryAppender.search("Transaction started", Level.ERROR)).hasSize(1);
         assertThat(memoryAppender.search("Transaction started", Level.TRACE)).isEmpty();
     }
 
     @Test
-    public void whenUsingPattern_thenReturnExpectedResult() {
+    void whenUsingPattern_thenReturnExpectedResult() {
         BusinessWorker worker = new BusinessWorker();
         worker.generateLogs("Order processed successfully for Order ID: 12345");
 
         Pattern orderPattern = Pattern.compile(".*Order ID: \\d{5}.*");
 
+        assertThat(memoryAppender.containsPattern(orderPattern, Level.DEBUG)).isTrue();
         assertThat(memoryAppender.containsPattern(orderPattern, Level.INFO)).isTrue();
         assertThat(memoryAppender.containsPattern(orderPattern, Level.WARN)).isTrue();
         assertThat(memoryAppender.containsPattern(orderPattern, Level.ERROR)).isTrue();
@@ -79,7 +78,7 @@ public class BusinessWorkerUnitTest {
     }
 
     @Test
-    public void whenUsingMultiplePatterns_thenReturnExpectedResult() {
+    void whenUsingMultiplePatterns_thenReturnExpectedResult() {
         BusinessWorker worker = new BusinessWorker();
         worker.generateLogs("User Login: username=user123, timestamp=2024-11-25T10:15:30");
 
@@ -88,7 +87,21 @@ public class BusinessWorkerUnitTest {
             Pattern.compile(".*timestamp=\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}.*")
         );
 
+        assertThat(memoryAppender.containsPatterns(patterns, Level.DEBUG)).isTrue();
         assertThat(memoryAppender.containsPatterns(patterns, Level.INFO)).isTrue();
         assertThat(memoryAppender.containsPatterns(patterns, Level.WARN)).isTrue();
+        assertThat(memoryAppender.containsPatterns(patterns, Level.ERROR)).isTrue();
+        assertThat(memoryAppender.containsPatterns(patterns, Level.TRACE)).isFalse();
     }
+
+    @Test
+    void whenLogsGenerated_thenCapturedOutputContainsLogs(CapturedOutput capturedOutput) {
+        String log = "Order processed successfully for Order ID: 12345.";
+        BusinessWorker worker = new BusinessWorker();
+        worker.generateLogs(log);
+
+        assertThat(capturedOutput.getOut()).contains(log);
+        assertThat(capturedOutput.getOut()).containsPattern(".*Order ID: \\d{5}.*");
+    }
+
 }
