@@ -13,11 +13,12 @@ public class SecurityConfig {
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http.authorizeHttpRequests(expressionInterceptUrlRegistry ->
-                        expressionInterceptUrlRegistry
-                                .anyRequest()
-                                .permitAll())
-                .csrf(AbstractHttpConfigurer::disable);
+        http
+          .authorizeHttpRequests(expressionInterceptUrlRegistry ->
+            expressionInterceptUrlRegistry
+              .anyRequest()
+              .permitAll())
+          .csrf(AbstractHttpConfigurer::disable);
         return http.build();
     }
 }
